Updated openstack/openstack Project: openstack/nova-specs  4d886bb746bd76ddb01b4a439c8192264cda9e05
RequestSpec object split instance_properties

To prepare for later uses of the RequestSpec object, and so that it can
be more explicit about what is available for use during scheduling, the
instance that has been shoved into instance_properties has been split
into the component pieces that are used by current scheduler filters.

vm_state and task_state are used as well, but only for the host state
consume_from_instance method and they are never different between
requests so we can leave them out of there.

image_ref is used for one filter but should be available in
requestspec.image.

Change-Id: I5744367ce395dac1d300a740f67c0a17668dd952