Updated openstack/openstack Project: openstack/tempest  f27e00917c191d0af74143cf76a2a8316487dcd9
Decouple Sahara templates test from vanilla plugin

The change introduces a list of template data for node templates and
clusters for each plugin. The first known plugin and its first known
version are used.

At least one default template is added for each plugin, so that all
the real deployment can be covered.

The tests are skipped if the plugins key in tempest configuration is empty;
they fail if no known plugin or no known version for known plugins are found
(which probably means that the test configuration should be updated).

The list must be maintained; on the other way, after some discussions,
the alternative solution - expose all the node group and cluster
templates parameters in tempest configuration - would have made the
testing more random and too dependent on the cloud operator in charge
of running tempest.
The list of plugins is not probably going to change massively for each
release, so it seems feasible.

Change-Id: I397616f7923e696082fe67da3ff9841f89e9a505