Updated openstack/openstack Project: openstack/requirements  11c700cd3896df2893df36cffd604a236af07351
Revert "Disable syncs to libs to unjam master requirements changes"

Now that requirements have been updated in master of libraries and
new releases of them tagged, the change to remove them from the
requirements sync can be reverted for now. We will work out a better
mechanism to make this tractable in the future.

This reverts commit 517f2174b025a4b7859e63308c3b58f4b4130e79.

Change-Id: I835b110b09d145682a87c069b6f9bbe86a1b1681