Updated openstack/openstack Project: openstack-infra/devstack-gate  a2e4844b32b99d47da31fa8cb4d87a3b61a345ea
Add errexit to setup & cleanup functions

It is too easy to introduce errors into the gate scripts and not
notice; for example see Icd853c670dfd9bd067f472203b38dd258d90e8ba.  To
find problems like this, set errexit for the setup & cleanup helper
functions.

To facilitate this more cleanly we add a little helper function that
creates the commands.

Dependent changes should have made the scripts errexit clean.
cleanup_host is left for another change to avoid too much churn.

Note setting errexit globally in the devstck-vm-gate-wrap.sh script is
a little more trouble than it is worth; since the testing commands are
expected to fail if there is a problem, we need to continue with the
script to clean-up and copy the logs out.

You might think, hey, we could get around that by doing some variation
of

  a_function && rc=$? || rc=$?
  cleanup_fn
  return $rc

but there is a small trap with that -- running as part of a "&&" or
"||" pipeline disables errexit for "a_function" -- thus by setting
errexit in the calling script, we effectively disable ourselves from
using it in functions we call that might fail!  This is a problem as
we're using bash functions as hooks all over the place.

Change-Id: Ia791b8bdb9fef5d5c41702eaa8b43c2e94b8499b