Updated openstack/openstack Project: openstack/cinder  f794e8f47f9a075b6264dd72d3f59c8153fa7d81
Add external genconfig calls

After moving to oslo.config we still were using
incubator config generator.  This was ok, but the
problem is we haven't been pulling config options
from the oslo libs.

This is a hack that just appends external lib calls
and appends those options to the sample file being built.

Change-Id: I2634b20ef4abd3bf7990f845d59ad3d208db234f
Closes-Bug: #1447380