Updated openstack/openstack Project: openstack/neutron  dfcfbff4dbdb83695392550d0877673aafc62e72
Add block name to switch config options for MLX plug-ins.

In the INI files, the switch_names option uses a dynamic
value to determine the block names for the switch options.
In order to create proper config option reference docs,
there needs to be an example block name for the switch
options.

Change-Id: Ic5bf6de02ba1b7d1bc90ee29a5a0570fb45b9956
Closes-Bug: #1442357