Updated openstack/openstack Project: openstack/glance  582f8563e866f167ae1de1a2309c1a1e2484442a
Metadef JSON files need to be updated

Since Juno, some new hardware properties have been added to Nova that
should be included in the Kilo versions of the metadef JSON files.

This bug, would add in new properties or objects found in Nova and update
properties or objects if the definitions have changed.
Closes-Bug: 1436877

Change-Id: I582827e666bece539674e4a191963ecf8ebf599f