Updated openstack/openstack Project: openstack/neutron  4e208ef120aa9e489b2861dda6e5f974f5e00ca2
ovs_lib: Fix a race between get_port_tag_dict and port removal

get_port_tag_dict() gets a list of ports using get_port_name_list()
and then queries the db again for ports in the list.
It fails if some of ports disappeared in between.
This change fixes it by ignoring "not exist" errors in the later query.

Closes-Bug: #1444797
Change-Id: Ic54b644bb1d72a4664b70f124863d17805c26fff