Updated openstack/openstack Project: openstack/tempest  ab94185bc54260cf107e19070ea8e71502b5cf8b
Rename an internal method to _helper_list()

helper_list() is just a helper method for "list_<resoruces>" of
telemetry_client and it is used in telemetry_client only.
telemetry_client will be migrated to tempest-lib and we need to
clarify whether each method is external or internal.
This patch renames the method for clarifying an internal method.

Partially implements blueprint consistent-service-method-names

Change-Id: I12922cf1ca968eb1d9d793879aa22921cc39a56b