Updated openstack/openstack Project: openstack/requirements  649de388d861d2a0ead94fcf3f66ecdebc1f926b
Bump eventlet version to 0.17.3

eventlet 0.17.3 is the first release fully compatible with Python 3 when
eventlet is used with monkey-patching. It is required to port OpenStack
to Python 3. For example, Oslo Messaging test suite now pass with
eventlet 0.17.3. Currently, eventlet is disabled on Python 3 in Oslo
Messaging, which prevents many applications to be ported to Python 3.

More information on the project to port OpenStack to Python 3:
https://wiki.openstack.org/wiki/Python3

Moreover, eventlet 0.17.2 includes a change to use a real logger
interface, so it makes possible to cleanup oslo.log (more information in
the bug #1440773).

Change-Id: I73e3056b5e8b9ce710c9c2d59fc5be8e03e28d2a
Partial-Bug: #1440773