Updated openstack/openstack Project: openstack/python-ironicclient  a68770bbb371cb7085aea2d4d2a2048b27a1f23b
Ensure *-show input uuid is not empty

Ironic *-show does not ensure that uuid are not empty.

This patch adds checks to the client to ensure that command line
arguments are not empty strings or only whitespace. It will raise
a command exception if *-show command line argument is an empty
string or only whitespace.

Change-Id: Ia34957922006da5ecbbfc7483040d67ac51f0ada
Closes-Bug: 1442436