Updated openstack/openstack Project: openstack-dev/grenade  8ee1af8f3b9fd55a990ff51409623ccd6b80e3e2
clean up upgrade scripts

The upgrade scripts have a *ton* of historical cruft in them. This is
a bulk clean up patch that does the following things:

 - ensures we source grenaderc before functions, we need this for
   devstack dir definitions.

 - moves xtraces until after sources to trim down irrelevant output

 - removes as many variables and library sources as possible to ensure
   that we include the minimum possible code in these scripts (for
   sanity and readability).

This skips the neutron upgrade script, much like in devstack, the
neutron code has so many extra variables it's not clear what's really
needed.

Change-Id: Idac9190fd33478a479f7c7dd5cb1be894bb03444