Updated openstack/openstack Project: openstack/tempest  214a6e440a45a986fe4d0aede3bcd8cd704e4ff5
Add logging of account allocation to accounts provider

This commit adds a log message to the locking accounts cred provider
to indicate when it has allocated a set of credentials successfully
from the accounts file. Right now when reading the tempest log file
during a run when using an accounts file there is no way to tell which
accounts where returned by the cred provider. This adds a info log msg
similar to what we log in isolated creds when it creates a new set of
creds. This should make it easier to trace through when credentials are
being allocated from the accounts file.

Change-Id: I82af7a3cba1e62cada94a9f4220666f1733e0575