Updated openstack/openstack Project: openstack/neutron-vpnaas  5c1d1b0180f642faf4d8d2341386d04926d4b288
Fix failures for integration tests

These unit tests were affected by the lack of callback registration.

Callbacks can be wiped out depending on the order in which subscriptions
and mocking of the callback registry manager takes place. By subscribing
during setUp, this ensures that the integration is in place so that
tests can pass.

Closes-bug: #1448245

Change-Id: I0dfa1ffa39403cbe00d2724fe21e6e22ef77e07d