Updated openstack/openstack Project: openstack/keystone  ec2df5897ec01763eecb23f062eb5118313cdc3e
Tests use Database fixture

The Database fixture already kept a reference to the engine,
and was already registering the cleanups that the tests were
registering again. All the fixture needed was a method to
recreate the tables.

Also, moved the cleanup in the Database fixture before create_all,
so that it happens even if create_all raises.

Change-Id: I9a65b15289f4c0342383fa1da58bede942ab53aa