Updated openstack/openstack Project: openstack/oslo.messaging  712dcc4ca4286d91414fe800fa5f59e5b34e6664
Enable eventlet dependency on Python 3

Bump eventlet to 0.17.3, the first release fully supporting Python 3
with monkey-patching.

Add aioeventlet and trollius dependencies for the aioeventlet
executor on Python 3.

This change enables tests of eventlet and aioeventlet executors on
Python 3.

Add futures to Python 3 dependencies even if it's not needed, it's
required to workaround a bug in tox.

Depends-on: I73e3056b5e8b9ce710c9c2d59fc5be8e03e28d2a
Change-Id: I0efae1c91c5d830156b867d7d21b5c0065094665