Updated openstack/openstack Project: openstack-infra/shade  0a3cd476437e08717f3498ff93b905198407a666
Do not cache unsteady state images

States that are expected to change should not be cached, as this will
just prolong any polling loops to detect the changes. This also makes
for less invalidations as we can depend on the fact that the cache will
only ever contain steady states.

Change-Id: Ie241db6311540725d24dcc6328f6f965820baf1b