Updated openstack/openstack Project: openstack/openstack-manuals  2dd540e688173c8eb23dd3f73403a83be7260616
[install-guide] Specify lock_path option

The lock_path option moved from the [DEFAULT] section to
the [oslo_concurrency] section in several services. However,
some distributions did not update their configuration files
to reflect this change. In these cases, we must explicitly
add this option in the correct location to prevent
deprecation warnings.

Update this patch to address additional services with this
issue.

Change-Id: I208223e60dd86f9e4c753b52a2faaf240209b674
Implements: blueprint installguide-kilo