Updated openstack/openstack Project: openstack/python-openstackclient  416d840dc4cb00026bac2512b259ce88a0e4a765
Reduce parameters to base class execute

Simplify the parameters so we are just passing a command string
to the execute command in the base class.  The string is exactly
the command we are going to run.  This will make debugging
easier and make it clearer what we are actually running.

Change-Id: I0425007e1849f31d692420e38544c55b1acb86c4