Updated openstack/openstack Project: openstack/taskflow  7edd1c07c5327c3cf42575d1c27c1a6344a79b2a
Add a conductor running example

Create an example which can be extended to create a simplistic
review checkout, tox running system which can be used to run
some set of actions on every review posted. This could be expanded
and connected into a gerrit pipeline to create a mini-jenkins
like trigger/build/result system.

Part of ongoing blueprint more-examples

Change-Id: I5cf1bf02eeddf897ac7f098f1d73377f262a267b