Updated openstack/openstack Project: openstack/manila  5458adb5d243e675b9765d76f176cdfdd756e333
Allow overriding the manila test regex

To make the script more flexible to include more or less tests,
Allow the test regex to be defined prior to calling to post_test_hook.sh.

This will also allow the various jenkins jobs to define the regex directly
instead of relying on their name.

Change-Id: Id64d761e7ad035a2e6d6351b04b9f224aa3b8e86