Updated openstack/openstack Project: openstack/horizon  d0a0976e78b9c9019cb3c093b71cc6026ed16b36
Removing unused variables in dashboard

Before we can enable a global check for unused variables, we need to clean
up the code first. This way, the gate won't barf when we try to enable the
global unused and undef jshint config.

This patch addresses the files in openstack dashboard only.

Change-Id: I4b0625ead27724ad7378ad906e2df26e00cc56ac
Partial-Bug: #1441794