Updated openstack/openstack Project: openstack/python-openstackclient  48b52a0d40013d65e85ebeadcfb80a06c431b6db
Raise exception if no session is created

If the clientmanager is unable to create a keystone session (due to
insufficient auth parameters or something else) then the exception
caused by this shouldn't be ignored, as was the case.

On the other hand, we don't want this behaviour in the case of the
'complete' command, so this is now properly detected.

Change-Id: If4f453d23cc87900cda752e9ffbcf41ded59e26f
Closes-Bug: #1444640