Updated openstack/openstack Project: openstack/tempest-lib  c307ffc525d896e2071319c8067aec805804f92b
Switch from using custom urlparse import logic to six

This commit removes the try except logic in auth.py to import urlparse
on either python2 or python3 to use six instead. Six already provides
an interface to use urlparse the same way on either python2 or python3
which is a bit cleaner.

Change-Id: I964586af4b2df63ae81b09ac1c3f1397b0250b80