Updated openstack/openstack Project: openstack/tooz  74c16fe88e4b60c8f9ca1dd9c9b8ca6b6f85ab55
Heartbeat on acquired locks copy

Instead of heartbeating on a collection that may
be changing size while iterating ensure that we only
iterate over a copy (that won't change size) during
iteration.

The python set() docs say that it is inherently
unpredictable to iterate on a set while its being mutated
so we should try to avoid that kind of case if we can.

Change-Id: I6f103a5528b3f62f1c8d9e501e8cadff4221ee40