Updated openstack/openstack Project: openstack/oslo.serialization  166726d545aa912e08a2466330296f9f9e13fe7e
Fixes wrong folder for listing tests

oslo.serialization had ./tests in its .testr.conf instead of just .,
which as a result didn't run the tests inside ./oslo_serialization.
This is quite bad, because what's important right now is the new
tests in the non-namespace folder. This patch fixes this, and allow
to run all tests.

Change-Id: Ic8d7409b214864d6350c76fd6d86c430a7380e8c