Updated openstack/openstack Project: openstack/tooz  0e67e15ed38d4f605380e67323265fa3b1dde88d
Avoid using a thread local token storage

When one thread is running the coordinator heartbeat()
periodically and another is using the lock(s) it should
be possible to correctly make this happen/work without
errors of the type of:

AttributeError: '_thread._local' object has no
                attribute 'token'

This happens because the coordinator heartbeat() method
calls into each current lock(s) heartbeat()
method (which then extends the lock via the locks
extend() method); when this is using thread local storage
this fails with the above error.

So to avoid this explicitly disallow using thread
local token storage by ensuring we pass in the right param
to the underlying redis lock.

Change-Id: I8e5592813577c3cf1cf9a67d0005d1e0ef180fca