Updated openstack/openstack Project: openstack/nova  bc6c891fd93dc91154af028d90003c611c8cab37
Avoid indexing into an empty list in getcallargs

If a method signature has no named arguments, either empty or
(*args, **kwargs), then inspect.getargspec() will return an empty list
for argnames.  In that case we should not try to compare against the
first element of that list.

Also updates the docstring for getcallargs to specify a limitation of
the method.

Change-Id: I68af587fe91310a68bb5c37b80d3f00f5d45dd16