Updated openstack/openstack Project: openstack/tempest  9c95882542acd47779066f46bd9b771dc06678f6
Switch all uses of StringIO to use it from six

As part of enabling python3 support in tempest we need to be able to
handle StringIO usage in both python2 and python3. Six provides a
compat layer for doing this, so this commit moves all uses of StringIO
and cStringIO to get it through six.

Change-Id: Ie6ac86b3b5ed2c307dc2cc41386da9e5ba1ee23e