Updated openstack/openstack Project: openstack/tempest  3c634d1419a5ea4600355e8b4180b08cf4e7615c
not all scenario tests should be in smoke

The decision to include all of scenario by default makes for an
extremely heavy test of tests in smoke, we should instead only include
specific scenario tests that expose some basic sanity
checking. Advanced scenario tests should be dropped from this list.

This change means we no longer have to exclude 'slow' tests, as we'll
just tag the right tests with smoke.

Change-Id: I34aa8aa9274e02bb68f62863d64c47ba9916f2e4