Updated openstack/openstack Project: openstack/heat-translator  66394b802c43cf104da62ebefeaf0e8ef3b3f9d1
Create a unit test for wordpress specification in TOSCA

Create a TOSCA-HOT translation unit test template and use that to
unit test TOSCA wordpress example.

Note, to implement this patch some changes had to be made to the
wordpress TOSCA file as translation of certain pieces were not fully
supported, as described in bug #1440247. As a result of the change in
wordpress TOSCA file some existing test cases had to be skipped as
well. It is important to re-enable the skipped code and revert the
changes made to wordpress TOSCA file as part of bug #1440247 patch.

Also, shell scripts that support mysql, web server, and wordpress
installation and configuration were also updated to make sure the
HOT translation works with the ubuntu-software-config-os-init image.

To address bug #1447800 handling of null dictionaries was also
improved.

Change-Id: I2710eabb0a74c4d75498ac2d0fe3e50ea50133ab
Closes-Bug: #1435545
Closes-Bug: #1447800
Partial-Bug: #1436420
Related-Bug: #1440247