Updated openstack/openstack Project: openstack/tripleo-incubator  8194f82f87e448c5ba01b8e5870640705ed5659f
create-nodes: allow multiple bridge names

This patch updates create-nodes so that it supports
configuring VMs which are attached to multiple baremetal
bridges. The name of the virtual machine (previously based
on the bridge) has been adjusted to concatonate all the
bridge names as a convention when naming virtual machines
with multiple bridges.

Change-Id: I5857629df7c8f7bc00c3a475a3e8ffc6f7e99c93