Updated openstack/openstack Project: openstack/tempest  693ed4f8c5ee762de937338eafbabf5bfaf4a11d
Switch all uses of urlparse to import from six

As part of enabling python3 support in tempest we need to be able to
handle urlparse usage in both python2 and python3. Six provies a compat
layer for doing this, so this commit moves all uses of urllparse to
get it through six.

Change-Id: Ic9c15d90fd3b8413fec4d01a070cc167f7536e62