Updated openstack/openstack Project: openstack/glance_store  ffe6e255508fdb2fe705878e9deb6d856e649f6b
Properly instantiate Forbidden exception

The Forbidden class (and its parent GlanceStoreException) do not take in
any arguments to the init method. Instead, since the message used when
os.unlink fails is different than the default message in the Forbidden
class, a keyword "message" argument needs to be used to properly
instantiate the class.

Change-Id: Iebe1891a61dc4874658065b727b126b72bcd3144
Closes-Bug: #1444663