Updated openstack/openstack Project: openstack/nova  244dfaf5f6bdfe51e2ffb8590c5f364168dfc624
Associate floating IPs with first v4 fixed IP if none specified

In the absence of a specified fixed address with which to associate a
floating IP, the first IPv4 address on the port should be associated.
Without the check for IPv4, IPv6 ports can be associated with a (IPv4)
floating IP, which is not supported.

Change-Id: Ib66a9109cc1c7999474daca5970d0af1f70886e4
Closes-Bug: 1437855