Updated openstack/openstack Project: openstack/nova  93610787de148c4bb01f2b9ee903c2d7e27e172a
VMware: Ensure compute_node.hypervisor_hostname is unique

The VMware driver currently identifies a hypervisor as:

  <mo_id>(<cluster name>)

Cluster name is not unique, and mo_id is only unique in the scope of a
single vCenter, so the above can produce duplicate names when managing
clusters with the same name in multiple vCenters. This is especially
likely to occur in lab setups.

This change alters the above format to the following, which uniquely
identifies a cluster globally:

  <mo_id>.<vCenter uuid>

Closes-bug: #1329261
Co-Authored-By: Kanagaraj Manickam <kanagaraj.manickam@hp.com>
Change-Id: I2f3b5d224cc653d0465598de0788116e71d1ca0d