Updated openstack/openstack Project: openstack/diskimage-builder  8b002cba09dc2be29439dcb7a86e4ffc09afc59e
Clean up targetcli ramdisk installation

The targetcli element was triggering a bunch of errors from dracut
when we installed all of Python.  It turns out this is because there
were filenames with spaces in the find output and the loop didn't
handle that properly.  This switches to a while loop that can
handle odd filenames.

Change-Id: Iacbf16f26f2bc9991840250dc8ae7990db54d811