Updated openstack/openstack Project: openstack/diskimage-builder  d2727fccf40a5d0277e8e0e7bad6c93021c98b88
Split the debootstrap functions into an element

debootstrap is not debian or ubuntu specific. We can make a debootstrap
element that knows how to do all of the things, and then a
debian-minimal and ubuntu-minimal image that use it. Finally, make
the debian element simply be a collection of the extra things we do to
make it look like a cloud-init based cloud image.

Change-Id: Iaf46c8e61bf1cac9a096cbfd75d6d6a9111b701e