Updated openstack/openstack Project: openstack/heat  c26160db0c5660ceaaf1498e688c7f703bd9d33b
Use list instead of iterator in `for` loop for modifiable collection 

six.iterkeys() returns iterator. Using this iterator while
removing entries from dict may raise RuntimeError. So we 
shouldn't use iterator in this place. Use list of `stack_id`
values instead.
For py2 we can use .keys(), but .keys() method returns `view`
in py3 and this will lead to the same error as with using
six.iterkeys(). Apply list() to the result given by .keys() to
produce the same behaviour for py2 and py3.
Also refactor tests to make sure that there are no problems
with iterators.


Co-Authored-By: Oleksii Chuprykov <ochuprykov@mirantis.com>
Change-Id: Ice43e18173b4b523d98703bca60819d0e62cfc91
Closes-Bug: #1448186