Updated openstack/openstack Project: openstack/barbican  597869880f186ce951809fe85d5d7d0610f35c4f
Fixed Bug for KMIP Secret Storage

The KMIP secret store was incorrectly storing secrets. In some cases
this resulted in extra information being stored with the keys and in
other cases the key storage would fail with a 500 internal server
error.

This patch fixes the KMIP secret store to correctly store secrets.

Change-Id: I94944a05776d366bd33d46ddb25f7129425405d0
Co-authored-by: Kaitlin Farr <Kaitlin.Farr@jhuapl.edu>
Closes-Bug: #1449234