Updated openstack/openstack Project: openstack/trove  ad008daebaf9da16f1864d6ab24430459ec31de0
Fix process attribute check in BackupRunner

Base 'BackupRunner' checks for the 'process' attribute on __exit__
and terminates the runner process if any.
It however always initializes the process to 'None' in the __init__.
That check is therefore always True.
It should change the condition to check for 'None' value instead.

I also made the 'run' method 'protected' to keep it consistent with the
'pre' and 'post' methods (and also the restore runner which
has the process handling method protected as well).

Change-Id: I6e98a47d33de7da95d884fb67b2cfab918183b3b
Closes-Bug: 1448279