Updated openstack/openstack Project: openstack/ironic  8dad33164b42a39493a695c1716f061c27e09278
Remove state transition: CLEANFAIL -> CLEANING

The state machine incorrectly models the transition from CLEANFAIL to
to CLEANING state. The spec [1] mentions that management of CLEANING
tasks should be handled in the same fashion as ZAPPING tasks. For
ZAPFAIL:
    Nodes that transition into ZAPFAIL will automatically enter
    maintenance mode, as failure to ZAP a machine usually indicates
    a hardware failure or something else that requires remote hands
    to fix.

Nodes in CLEANFAIL should do a similar thing as ZAPFAIL; they are
put in maintenance mode, but shouldn't be able to transition directly
to CLEANING.

This patch also clarifies the documentation for the transition
CLEANFAIL -> MANAGEABLE.

[1] http://specs.openstack.org/openstack/ironic-specs/specs/kilo/new-ironic-state-machine.html#proposed-change

Change-Id: Ieda2fcd78332a063c51685cd01bb824014f0f4a8
Closes-Bug: #1446758