Updated openstack/openstack Project: openstack/sahara  bbb390f547a415518a976863228262999ed7675c
Date format set to be correct utc date

As it turned out the format of the date used in Sahara may not be
understood by som JavaScript engines which prevents the dates to be
displayed correctly in UI. The milliseconds part of the time is also
removed as it has never been necessary to access the dates with that
accuracy. However the presence or sometime absence of the .ms suffix
makes it harder to handle dates.

Closes-Bug: #1443343
Change-Id: I1f4cc6dc443605ba632c1458bba8bd47a0e1a01f