Updated openstack/openstack Project: openstack/manila  7fa41f854d5b8a9f3b67fcb18566550795e186cd
Add doc with basic deployment steps

This doc describes steps that should be performed to install simple Manila
configuration in existing cloud.
It has installation-specific steps that can differ from deployment to
deployment.

Following steps are described here:
- Installation of Manila binaries
- Installation of Manila client
- Registration in Keystone
- Preparation of external files (configs, etc...)
- Basic configuration of Manila
- Setup of DB
- Running services of Manila
- Creation of pilot share

Not covered:
- network plugins configuration
- share drivers-specific configurations to all drivers except Generic driver,
and this one is covered in only small use case.

Change-Id: Ie68d861411358d227b6f7cc3487dffd1e0838e4d