Updated openstack/openstack Project: openstack-infra/storyboard-webclient  fcd1561c17a4810999a6097086c807862ae71ff1
Updated dependencies

An update in a fuzzy-matched dependency of angular-eslint is causing
our builds to break. Since we're not actually using it, I've removed
it.

Also, it appears that transient dependencies in grunt have been
removed, and thus grunt-connect-proxy started to break. This updates
that library and resolves the error.

Change-Id: I9da36cca23f3e900302b24c35b991f7790a681cb