Updated openstack/openstack Project: openstack-dev/devstack  6d45e29bbb97ad338b413ec31b635c3c4364df87
Set DevStack to dual stack by default in Kilo+

This patch sets DevStack to run in dual stack networking, with both IPv4
and IPv6 networking configured. This change is required for dual stack
testing at the gate.

A different patch was created against devstack-gate to make this the
default, but the Juno branch of Neutron is missing required fixes to the
L3 agent that are present in Kilo. This was the suggested alternative.

Related-change: I3d416275f77913769b98e77f7e47bed17fc4d1cc

Co-Authored-By: Henry Gessau <gessau@cisco.com>
Co-Authored-By: Andrew Boik <dboik@cisco.com>

Depends-On: Ib66a9109cc1c7999474daca5970d0af1f70886e4
Depends-On: I0f9ea98cb84aa72cb1505fb9ff8ac61561cc1376
Depends-On: I85fe68782bc54f28f3e14aa4a1d042cb15959dac
Depends-On: I9395834f673038dc23b25eaeefe14895fe154e0e

Change-Id: If0e0b818355e4cb1338f7fa72af5e81e24361574