Updated openstack/openstack Project: openstack/governance  d5d8d404ecf2ff8aed414cb0c812e8ac4dc15000
Clarify that meetings use official IRC channels

The 2014-12-02 OpenStack project structure reform resolution
indicates that using "#openstack-meeting channels for their team
meetings" is a one-of-us criterion for being considered an OpenStack
project. Clarify this point in the new project application
requirements. Also spell out that this means the meetings are held
in public for all to participate, and the logs are published
somewhere so that anyone can review them later.

Change-Id: Ia34e9bb1d2db3094f61968b2bb9a7849f2bc7fd9