Updated openstack/openstack Project: openstack/tempest  9f5de85ca0e36017360d9492a16939be155ab814
Switch all uses of ConfigParser to import from six

As part of enabling python3 support in tempest we need to be able to
handle ConfigParser usage in both python2 and python3. Six provies a
compat layer for doing this, so this commit moves all uses of
ConfigParser to get it through six.

Change-Id: Ie4a788dc9b9151557e9f197fb48f3225c39de89f